+++ Forge/changelog.md
@@ -1,2 +1,2 @@
-1.2.1:
-* Change corn to use cross model instead of crop model
+1.2.2:
+* Fix corn growth stages and drops